major commit

+++ app/Http/Controllers/PostsController.php
@@ -43,8 +43,9 @@ class PostsController extends Controller
         $single_element=new POST();
         $single_element->title=$request->title;
         $single_element->body=Purifier::clean($request->body);
+        $single_element->user_id=auth()->user()->id;
         $single_element->save();
-        return redirect('/posts')->with('success','Post Created');
+        return redirect('/dashboard')->with('success','Post Created');
     }
 
     /**
@@ -68,6 +69,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $single=POST::find($id);
+        $userid=auth()->user()->id;
+        if($single->user_id != $userid){
+            return redirect('/dashboard')->with('error','Unauthorized page');
+        }
         return view('posts/edit')->with('single',$single);
     }
 
@@ -88,7 +93,7 @@ class PostsController extends Controller
         $single_element->title=$request->title;
         $single_element->body=Purifier::clean($request->body);
         $single_element->save();
-        return redirect('/posts')->with('success','Post Updated');
+        return redirect('/dashboard')->with('success','Post Updated');
     }
 
     /**
@@ -100,7 +105,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post=Post::find($id);
+        $userid=auth()->user()->id;
+        if($post->user_id != $userid){
+            return redirect('/dashboard')->with('error','Unauthorized page');
+        }
         $post->delete();
-        return redirect('/posts')->with('success','Post Deleted');
+        return redirect('/dashboard')->with('success','Post Deleted');
     }
 }
